refactor(#3094): extract the inline test mods of generation_merge.rs + partition_shadow.rs (#1116/#1135)
Both files are at/over the campsite-rule threshold with zero headroom
(generation_merge.rs 831/800, partition_shadow.rs 800/800), so the multi-gen
presence fix cannot add a line to either without tripping the file-size
ratchet. Move each file's '#[cfg(test)] mod tests { .. }' VERBATIM into a
sibling '*_tests.rs' included via '#[cfg(test)] #[path = ".."] mod tests;',
so 'super' still resolves to the same module and every path is unchanged.

  generation_merge.rs   831 -> 684
  partition_shadow.rs   800 -> 396

Behaviour-preserving by construction; verified by diffing the '--lib --list'
test-name set for both modules before and after: IDENTICAL (23 names).
